Quantia/Jobs/SentimentAnalysisJob/requirements.txt
torch>=2.2
transformers==4.41.0
sentencepiece
sentence-transformers==2.7.0
keybert>=0.7.0
scikit-learn>=1.4
tiktoken>=0.6.0
spacy==3.7.3
spacy-langdetect
langdetect
optimum==1.20.0
onnxruntime
praw==7.7.1
openai==1.90.0
python-dotenv>=1.0
backoff==2.2.1
tqdm>=4.66
numpy>=1.26
pandas>=2.2
